requirements.txt
Django==1.11.6
django-ipware==1.1.6
django-markdown==0.8.4
django-markdown-deux==1.0.5
django-rest-framework==0.1.0
djangorestframework==3.7.1
easy-thumbnails==2.4.2
Markdown==2.6.9
markdown2==2.3.4
olefile==0.44
Pillow==4.3.0
pytz==2017.2
